Change AG Grid View, Add Cinema data view, fix null image file path, Add showOnTop

+++ db-world-backend/src/main/java/com/db/dbworld/services/Impl/LoginDataServiceImpl.java
@@ -46,7 +46,7 @@ public class LoginDataServiceImpl implements LoginDataService {
     }
 
     @Override
-    public Integer totalNumberOfLogin(Long userId) {
+    public Long totalNumberOfLogin(Long userId) {
         return this.loginDataRepository.totalNumberOfLogin(userId);
     }
 }
