+++ db-world-backend/src/main/java/com/db/dbworld/services/LoginDataService.java
@@ -5,5 +5,5 @@ import com.db.dbworld.payloads.user.LoginDataDto;
 
 public interface LoginDataService {
     LoginDataDto addAgentByUserId(String agent, Long userId);
-    Integer totalNumberOfLogin(Long userId);
+    Long totalNumberOfLogin(Long userId);
 }
